Bug fix - was using wrong GL blending mode, that resulted in the wrong appearance for images with alpha<1.
git-svn-id: http://svn.openimageio.org/oiio/trunk@353 31904243-a364-42ac-a3db-9305396518ba

+++ src/iv/imageviewer.cpp
@@ -500,7 +500,7 @@ ImageViewer::addRecentFile (const std::string &name)
 {
     removeRecentFile (name);
     m_recent_files.insert (m_recent_files.begin(), name);
-    if (m_recent_files.size() > MaxRecentFiles)
+    if (m_recent_files.size() > (size_t)MaxRecentFiles)
         m_recent_files.resize (MaxRecentFiles);
 }
